+++ src/GetSchemaPath.cpp
@@ -1,4 +1,4 @@
-#include "UpdateSchemas.h"
+#include "GetSchemaPath.h"
 #include <boost/dll.hpp>
 #include <boost/filesystem/path.hpp>
