+++ test/RequestHandlerTest.cpp
@@ -1,10 +1,10 @@
+#include "../src/RequestHandler.h"
 #include "../src/CustomExceptions.h"
 #include "../src/FlatbuffersTranslator.h"
+#include "../src/GetSchemaPath.h"
 #include "../src/Kafka/ConsumerInterface.h"
 #include "../src/Kafka/FakeConsumer.h"
 #include "../src/Kafka/FakeProducer.h"
-#include "../src/RequestHandler.h"
-#include "../src/UpdateSchemas.h"
 #include <boost/filesystem.hpp>
 #include <flatbuffers/idl.h>
 #include <gtest/gtest.h>
